test: refactor fixtures into submodules
