+++ tests/_fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Domain-grouped pytest fixtures, re-exported by tests/conftest.py."""
